REF: split policy_logger into separate module
